container/requirements.txt
# Core dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Async support
asyncio==3.4.3
aiohttp==3.9.1

# LLM provider integrations
openai==1.6.1
anthropic==0.8.1
httpx==0.26.0

# MCP protocol support
mcp==0.1.0  # Model Context Protocol library
websockets==12.0

# Configuration and serialization
pydantic==2.5.2
pydantic-settings==2.1.0
PyYAML==6.0.1
toml==0.10.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-grpc==0.8.0
black==23.12.0
isort==5.13.2
mypy==1.8.0

# Optional: Local model support
# ollama-python==0.1.7  # Uncomment if using Ollama